requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.9.0
distro-info==1.1+ubuntu0.2
docstring-parser==0.15
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==1.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
instructor==0.6.8
jeepney==0.7.1
kafka-python==2.0.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.5
netifaces==0.11.0
oauthlib==3.2.0
openai==1.35.14
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
PyYAML==5.4.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
SecretStorage==3.3.1
shell_gpt==1.4.3
six==1.16.0
sniffio==1.3.1
systemd-python==234
tenacity==8.5.0
tqdm==4.66.4
tweepy==4.14.0
typer==0.9.4
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
yarl==1.9.4
zipp==1.0.0
